==> HotelWise/00.DB.FIX.Amenities.ipynb <==
import pandas as pd
import random
import requests
import os

# DB FIX Amenities

## Agregar Datos Faltantes al Dataset

MONTAJE DE ARCHIVOS:

* Dataframe Hoteles

hoteles_dataset_content = "hoteles_unificado.xlsx"

if not os.path.exists(hoteles_dataset_content):
    url = "https://github.com/HotelWise/HotelWise/raw/Reviews/hospitales_unificado.xlsx"
    response = requests.get(url)
    with open(hoteles_dataset_content, 'wb') as f:
        f.write(response.content)
hoteles_dataset = pd.read_excel(hoteles_dataset_content)

hoteles_dataset.head(2)

print(hoteles_dataset.columns)

#Convertimos encabezados a mayúsculas
hoteles_dataset.columns = hoteles_dataset.columns.str.upper()
hoteles_dataset.head(2)

**Crear una lista de amenidades con el DB original**

import ast

# Crear una lista vacía para almacenar las amenidades sin repetir
amenidades_sin_repetir = []

# Iterar sobre cada celda de la columna 'AMENIDADES' y agregar las amenidades a la lista sin repetir
for amenidades_str in hoteles_dataset['AMENIDADES']:
    amenidades_list = ast.literal_eval(amenidades_str)  # Convertir la cadena a lista
    for amenidad in amenidades_list:
        if amenidad not in amenidades_sin_repetir:
            amenidades_sin_repetir.append(amenidad)

# Imprimir la lista de amenidades sin repetir
print(amenidades_sin_repetir)

# Crear una lista vacía para almacenar las amenidades sin repetir
amenidades_sin_repetir = []

# Iterar sobre cada celda de la columna 'amenidades' y agregar las amenidades a la lista sin repetir
for amenidades_str in hoteles_dataset['AMENIDADES']:
    # Convertir la cadena de amenidades a una lista usando eval
    amenidades_list = eval(amenidades_str)
    
    for amenidad in amenidades_list:
        if amenidad not in amenidades_sin_repetir:
            amenidades_sin_repetir.append(amenidad)

# Imprimir la lista de amenidades sin repetir
print(amenidades_sin_repetir)

## Agregado de **Amenities**

**Rellenamos Amenities hasta tener el df de GCP completo** 

# Lista de 30 amenidades
amenities_list = [
    'Pool', 'Gym', 'Restaurant', 'Bar', 'Spa',
    'Free Wi-Fi', 'Complimentary Breakfast', 'Room Service', 'Free Parking', '24-Hour Reception',
    'Business Center', 'Conference Room', 'Laundry Service', 'Daily Housekeeping', 'Air Conditioning',
    'Heating', 'Jacuzzi', 'Sauna', 'Hot Tub', 'Airport Shuttle Service',
    'Concierge Service', 'Accessible Accommodations', 'Family Rooms', 'Terrace or Garden',
    'Bicycle Rental Service', 'Beauty Salon', 'Luggage Storage', 'Children\'s Play Area',
    'Multilingual Staff', 'Soundproof Rooms'
]

# Lista extendida + 30 amenidades
amenities_list.extend([
    'Free Fitness Classes', 'Pet-Friendly', 'Indoor Pool', 'Outdoor Pool', 'Tennis Courts',
    'Basketball Court', 'Volleyball Court', 'Yoga Classes', 'Fitness Center', 'Library',
    'Game Room', 'Billiards', 'Ping Pong Table', 'Movie Theater', 'Karaoke Room',
    'Live Entertainment', 'Gift Shop', 'Mini Market', 'ATM', 'Currency Exchange',
    'Car Rental', 'Airport Shuttle', 'Train Station Shuttle', 'Beach Access', 'Water Sports',
    'Golf Course', 'Ski Resort Shuttle', 'Ski Equipment Rental', 'Hiking Trails', 'Bike Trails'
])
# Definir función para seleccionar 5 amenidades al azar sin repetición
def select_random_amenities():
    return random.sample(amenities_list, k=5)

# Llenar la columna 'amenities' con 5 amenidades aleatorias
hoteles_dataset['AMENITIES'] = hoteles_dataset.apply(lambda x: select_random_amenities(), axis=1)

hoteles_dataset.head(2)

###  Guardamos el df con las columnas necesarias para el análisis posterior

hoteles_para_guardar=hoteles_dataset[['NAME','LATITUDE','LONGITUDE','CITY','COUNTY', 'STATE', 'TEXT','AVG_RATING', 'AMENITIES']]

hoteles_para_guardar.head(2)

# Guardar DataFrame en un archivo Parquet
hoteles_para_guardar.to_parquet('hoteles_unificado.parquet')
