==> HotelWise/00.DB.FIX.Crime.ipynb <==
import pandas as pd
import random

# from google.colab import drive

# DB FIX Crime

## Agregar Datos Faltantes al Dataset

MONTAJE DE ARCHIVOS GOOGLE DRIVE:

# drive.mount('/content/drive')

* Dataframe Hoteles

# hoteles_dataset_content = '/content/drive/MyDrive/Colab Notebooks/HotelWiseML/hoteles_unificado.parquet'
# hoteles_dataset = pd.read_parquet(hoteles_dataset_content)

* Dataframe Seguridad

# crime_df_content = '/content/drive/MyDrive/Colab Notebooks/HotelWiseML/CrimeInTheUSA.parquet'
# crime_df_dataset = pd.read_parquet(crime_df_content)

MONTAJE DE ARCHIVOS LOCAL:

* Dataframe Hoteles

hoteles_dataset_content = 'hoteles_unificado.parquet'
hoteles_dataset = pd.read_parquet(hoteles_dataset_content)

* Dataframe Seguridad

crime_df_content = "Crime_In_The_USA/CrimeInTheUSA.parquet"
crime_df_dataset = pd.read_parquet(crime_df_content)

hoteles_dataset.head(2)

crime_df_dataset.head(2)

## Combinación db Hoteles con db Seguridad

# Realizar la fusión
hoteles_final = pd.merge(hoteles_dataset, crime_df_dataset[['STATE', 'CITY', 'CRIME RATE']], 
                         left_on=[hoteles_dataset['STATE'].str.lower(), hoteles_dataset['CITY'].str.lower()], 
                         right_on=[crime_df_dataset['STATE'].str.lower(), crime_df_dataset['CITY'].str.lower()], 
                         how='left')

# Eliminar las columnas adicionales
hoteles_final.drop(['key_0', 'key_1', 'CITY_y', 'STATE_y'], axis=1, inplace=True)
hoteles_final.rename(columns={'CITY_x': 'CITY', 'STATE_x': 'STATE'}, inplace=True)


hoteles_final.head(2)

###  Guardamos el df con las columnas necesarias para el análisis posterior

hoteles_para_guardar=hoteles_final[['NAME','LATITUDE','LONGITUDE','CITY','COUNTY', 'STATE', 'TEXT','AVG_RATING','CRIME RATE', 'AMENITIES']]

hoteles_para_guardar.rename(columns={'TEXT': 'REVIEWS','CRIME RATE':'CRIME_RATE'}, inplace=True)


hoteles_para_guardar.head(2)

len(hoteles_para_guardar)

celdas_vacias = hoteles_para_guardar['CRIME_RATE'].isna().sum()
print("Número de celdas vacías en la columna CRIME_RATE:", celdas_vacias)

# Guardar DataFrame en un archivo Parquet
hoteles_para_guardar.to_parquet('hoteles_unificado_final.parquet')
