==> HotelWise/01.NLP.01.NLTK.ipynb <==
!pip install contractions

import re
import pandas as pd
import nltk
from nltk.tokenize import word_tokenize
from nltk.corpus import stopwords
from nltk.stem import WordNetLemmatizer
from nltk.sentiment import SentimentIntensityAnalyzer
from contractions import fix
from google.colab import drive

# Descargar recursos (ejecutar una vez)
nltk.download('punkt')
nltk.download('stopwords')
nltk.download('wordnet')
nltk.download('vader_lexicon')

# NLP

## Análisis de Sentimiento con la Biblioteca NLTK y Diccionario Vader

MONTAJE DE ARCHIVOS GOOGLE DRIVE:

drive.mount('/content/drive')

user_reviews_content = '/content/drive/MyDrive/Colab Notebooks/HotelWiseML/hoteles_unificado.parquet'
user_reviews_dataset = pd.read_parquet(user_reviews_content)

MONTAJE DE ARCHIVOS LOCAL:

user_reviews_content = 'hoteles_unificado.parquet'
user_reviews_dataset = pd.read_parquet(user_reviews_content)

### Tokenizado, Lematizado y Stopwords

def preprocess_text(text):
    if not text or pd.isnull(text):  # Verificar si el texto está vacío o es nulo
        return ''

    # Eliminar emojis y caracteres especiales
    # Se pueden tokenizar los emojis con metodos mas robustos
    # que ayuden a manejarlos, pero se elije eliminarlos
    text = re.sub(r'[^\w\s]', '', text)

    # Arreglo de contracciones
    try:
        text = fix(text)  # Intentar usar la función fix()
    except Exception as e:
        print(f"Error al aplicar fix(): {e}")

    # Tokenización
    tokens = word_tokenize(text)

    # Palabras clave adicionales, positivas y negativas para mejorar el análisis
    negative_hotel_words = ['dirty', 'uncomfortable', 'noisy', 'smelly', 'outdated',
    'small', 'unfriendly', 'expensive', 'overpriced', 'unhygienic',
    'unsafe', 'crowded', 'inefficient', 'unorganized', 'rude',
    'disappointing', 'terrible', 'unreliable', 'dull', 'unresponsive',
    'unpleasant', 'inattentive', 'unsanitary', 'uninviting', 'dilapidated',
    'neglected', 'inconvenient', 'unaccommodating', 'problematic'
]

    positive_hotel_words = ['clean', 'comfortable', 'quiet', 'pleasant', 'modern',
    'spacious', 'friendly', 'affordable', 'luxurious', 'inviting',
    'safe', 'relaxing', 'efficient', 'organized', 'welcoming',
    'satisfying', 'excellent', 'reliable', 'enjoyable', 'responsive',
    'beautiful', 'attentive', 'sanitary', 'inviting', 'well-maintained',
    'cared-for', 'convenient', 'accommodating', 'problem-free', 'stellar'
]


    # Obtener stopwords y agregar palabras con connotación positiva y negativa
    stop_words = set(stopwords.words('english'))

    for word in negative_hotel_words:
        stop_words.discard(word)
    for word in positive_hotel_words:
        stop_words.discard(word)

    # Eliminación de stopwords y puntuación no necesarios
    tokens = [word for word in tokens if word.lower() not in stop_words and word.isalpha()]

    # Lematización
    lemmatizer = WordNetLemmatizer()
    tokens = [lemmatizer.lemmatize(word) for word in tokens]

    return ' '.join(tokens)
user_reviews_dataset['preprocess_text'] = user_reviews_dataset['reviews'].apply(preprocess_text)

### Función de Análisis de Sentimiento

# Función para analizar el sentimiento con VADER
def analizar_sentimiento(texto):
    analyzer = SentimentIntensityAnalyzer()
    scores = analyzer.polarity_scores(texto)
    compound_score = scores['compound']
    if compound_score >= 0.05:
        return 1 # Review Positiva
    else:
        return 0 # Review Negativa

user_reviews_dataset['sentiment_analysis'] = user_reviews_dataset['preprocess_text'].apply(analizar_sentimiento)

## Revisión de 'sentiment_analysis' vs. 'review'

A continuación vemos los resultados que se obtuvieron del análisis de sentimiento y comparamos con la columna 'review'.

filas_neg = user_reviews_dataset[user_reviews_dataset['sentiment_analysis'] == 0]
filas_pos = user_reviews_dataset[user_reviews_dataset['sentiment_analysis'] == 1]

pd.set_option('display.max_colwidth', None)

- Comparamos las filas negativas con la review

filas_neg[['reviews','sentiment_analysis']].head()

cantidad_ceros = (filas_neg['sentiment_analysis'] == 0).sum()

print("Cantidad de reviews negativas:", cantidad_ceros)

- Comparamos las filas positivas con la review

filas_pos[['reviews','sentiment_analysis']].head()

cantidad_unos = (filas_pos['sentiment_analysis'] == 1).sum()

print("Cantidad de reviews positivas:", cantidad_unos)

### Reordenamiento del dataset

user_reviews_dataset_Final=user_reviews_dataset[['name','latitude','longitude','city','county','avg_rating','security','sentiment_analysis']]

user_reviews_dataset_Final.head()

# Agrupar por 'name' y calcular la suma de 'avg_rating' y 'sentiment_analysis', conservando las otras columnas
summarized_data = user_reviews_dataset_Final.groupby('name').agg({
    'latitude': 'first',
    'longitude':'first',
    'county': 'first',
    'city': 'first',
    'avg_rating': 'sum',
    'sentiment_analysis': 'sum',
    'security':'sum',

})
summarized_data = summarized_data.reset_index()
summarized_data.head()

## Almacenamiento de Dataframe en Parquet

GUARDADO DE DATAFRAME EN GOOGLE DRIVE:

summarized_data.to_parquet('/content/drive/MyDrive/Colab Notebooks/HotelWiseML/Hoteles.NLP.01.NLTK.parquet')

GUARDADO DE DATAFRAME EN LOCAL:


summarized_data.to_parquet('Hoteles.NLP.01.NLTK.parquet')
