==> HotelWise/01.NLP.03.TensorFlow.ipynb <==
import pandas as pd
import numpy as np
import tensorflow as tf
from tensorflow.keras.preprocessing.text import Tokenizer
from tensorflow.keras.preprocessing.sequence import pad_sequences
from sklearn.model_selection import train_test_split
from google.colab import drive

# NLP

## Análisis de Sentimiento con las Bibliotecas Tensorflow, keras y sklearn, por arquitectura RNN Bidireccional (Red Neuronal Recurrente Bidireccional)

MONTAJE DE ARCHIVOS GOOGLE DRIVE:

drive.mount('/content/drive')

user_reviews_content = '/content/drive/MyDrive/Colab Notebooks/HotelWiseML/hoteles_unificado.parquet'
user_reviews_dataset = pd.read_parquet(user_reviews_content)

MONTAJE DE ARCHIVOS LOCAL:

user_reviews_content = 'hoteles_unificado.parquet'
user_reviews_dataset = pd.read_parquet(user_reviews_content)

### Tokenizado, Secuenciación y Armado

# Preprocesamiento de datos
reviews = user_reviews_dataset['reviews'].values

positive_keywords = ['clean', 'comfortable', 'quiet', 'pleasant', 'modern',
    'spacious', 'friendly', 'affordable', 'luxurious', 'inviting',
    'safe', 'relaxing', 'efficient', 'organized', 'welcoming',
    'satisfying', 'excellent', 'reliable', 'enjoyable', 'responsive',
    'beautiful', 'attentive', 'sanitary', 'inviting', 'well-maintained',
    'cared-for', 'convenient', 'accommodating', 'problem-free', 'stellar'
]

negative_keywords = ['dirty', 'uncomfortable', 'noisy', 'smelly', 'outdated',
    'small', 'unfriendly', 'expensive', 'overpriced', 'unhygienic',
    'unsafe', 'crowded', 'inefficient', 'unorganized', 'rude',
    'disappointing', 'terrible', 'unreliable', 'dull', 'unresponsive',
    'unpleasant', 'inattentive', 'unsanitary', 'uninviting', 'dilapidated',
    'neglected', 'inconvenient', 'unaccommodating', 'problematic'
]

# Función para determinar la etiqueta basada en las palabras clave
def determine_label(review):
    for word in positive_keywords:
        if word in review.lower():
            return 1
    for word in negative_keywords:
        if word in review.lower():
            return 0
    return 0  # Si no se encuentra ninguna palabra clave, se etiqueta como 'otro'

labels = [determine_label(review) for review in reviews]


# Tokenización y secuenciación
tokenizer = Tokenizer(num_words=10000, oov_token='<OOV>')
tokenizer.fit_on_texts(reviews)
sequences = tokenizer.texts_to_sequences(reviews)
padded_sequences = pad_sequences(sequences, maxlen=100, truncating='post')

# División de datos en conjuntos de entrenamiento y prueba
X_train, X_test, y_train, y_test = train_test_split(
    padded_sequences, labels, test_size=0.2, random_state=42)

### Construccion del Modelo de Red Neuronal Recurrente de Capa Bidireccional

# Construcción del modelo de análisis de sentimientos
model = tf.keras.Sequential([
    tf.keras.layers.Embedding(10000, 16, input_length=100),
    tf.keras.layers.Bidirectional(tf.keras.layers.LSTM(32, return_sequences=True)),
    tf.keras.layers.Flatten(),
    tf.keras.layers.Dense(24, activation='relu'),
    tf.keras.layers.Dense(1, activation='sigmoid')
    ])

### Compilado, Entrenamiento y Evaluacion del Modelo

# Convertir las listas de etiquetas en matrices NumPy
y_train = np.array(y_train)
y_test = np.array(y_test)

model.compile(loss='binary_crossentropy',
              optimizer='adam', metrics=['accuracy'])
model.fit(X_train, y_train, epochs=10, validation_data=(X_test, y_test))

# Evaluación del modelo
loss, accuracy = model.evaluate(X_test, y_test)
print("Accuracy: {:.2f}%".format(accuracy * 100))

#### Aplicado del Modelo al Dataset

# 1. Preprocesar los nuevos datos
nuevas_reviews = user_reviews_dataset['reviews'].values
nuevas_sequences = tokenizer.texts_to_sequences(nuevas_reviews)
nuevas_padded_sequences = pad_sequences(nuevas_sequences, maxlen=100, truncating='post')

# 3. Generar predicciones sobre los nuevos datos
nuevas_predicciones = model.predict(nuevas_padded_sequences)

# 4. Interpretar las predicciones (por ejemplo, etiquetarlas como positivas o negativas)
nuevos_sentimientos = [1 if pred > 0.5 else 0 for pred in nuevas_predicciones]

# Agregar los sentimientos predichos como una nueva columna en el DataFrame de nuevos datos
user_reviews_dataset['sentiment_analysis'] = nuevos_sentimientos

user_reviews_dataset.head()

### Implementacion en la DB

### Revisión de 'sentiment_analysis' vs. 'review'

A continuación vemos los resultados que se obtuvieron del análisis de sentimiento y comparamos con la columna 'review'.

filas_neg = user_reviews_dataset[user_reviews_dataset['sentiment_analysis'] == 0]
filas_pos = user_reviews_dataset[user_reviews_dataset['sentiment_analysis'] == 1]

pd.set_option('display.max_colwidth', None)

- Comparamos las filas negativas con la review

filas_neg[['reviews','sentiment_analysis']].head()

cantidad_ceros = (filas_neg['sentiment_analysis'] == 0).sum()

print("Cantidad de reviews negativas:", cantidad_ceros)

- Comparamos las filas positivas con la review

filas_pos[['reviews','sentiment_analysis']].head()

cantidad_unos = (filas_pos['sentiment_analysis'] == 1).sum()

print("Cantidad de reviews positivas:", cantidad_unos)

### Reordenamiento del dataset

user_reviews_dataset_Final=user_reviews_dataset[['name','latitude','longitude','city','county','avg_rating','security','sentiment_analysis']]

user_reviews_dataset_Final.head()

# Agrupar por 'name' y calcular la suma de 'avg_rating' y 'sentiment_analysis', conservando las otras columnas
summarized_data = user_reviews_dataset_Final.groupby('name').agg({
    'latitude': 'first',
    'longitude':'first',
    'county': 'first',
    'city': 'first',
    'avg_rating': 'sum',
    'sentiment_analysis': 'sum',
    'security':'sum',

})
summarized_data = summarized_data.reset_index()
summarized_data.head()

# Guardar DataFrame en un archivo Parquet
summarized_data.to_parquet('Hoteles.NLP.03.TensorFlowBidireccional.parquet')

## Almacenamiento de Dataframe en Parquet

GUARDADO DE DATAFRAME EN GOOGLE DRIVE:

summarized_data.to_parquet('/content/drive/MyDrive/Colab Notebooks/HotelWiseML/Hoteles.NLP.03.TensorFlow.parquet')

GUARDADO DE DATAFRAME EN LOCAL:


summarized_data.to_parquet('Hoteles.NLP.03.TensorFlow.parquet')
